+++ assets/javascripts/kitten/components/carousel/tab-bar-carousel/index.js
@@ -5,7 +5,7 @@ import { Deprecated } from '../../helpers/utils/deprecated'
 export const NavTabCarousel = props => {
   const message = `DEPRECATED:
     Import \`NavTabCarousel\` from \`…/nav-tab-carousel\`
-    instead of `…/tab-bar-carousel.`
+    instead of \`…/tab-bar-carousel.`
 
   return (
     <Deprecated warningMessage={message}>
