+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@kisskissbankbank/kitten",
-  "version": "1.10.0",
+  "version": "1.10.1",
   "license": "MIT",
   "description": "Front-end components library",
   "main": "src",
