+++ src/components/carousel/nav-tab-carousel/components/prev.js
@@ -7,7 +7,7 @@ var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefau
 Object.defineProperty(exports, "__esModule", {
   value: true
 });
-exports.Next = void 0;
+exports.Prev = void 0;
 
 var _react = _interopRequireDefault(require("react"));
 
@@ -19,10 +19,10 @@ var _typographyConfig = _interopRequireDefault(require("./../../../../constants/
 
 var _typography = require("./../../../../helpers/utils/typography");
 
-var RightArrowContainerStyles = _styledComponents.default.div.withConfig({
-  displayName: "next__RightArrowContainerStyles",
-  componentId: "bcawu0-0"
-})(["color:#fff;", ";align-items:center;display:flex;transition:transform 0.3s ease-in-out;&:hover{transform:translate(", ");cursor:pointer;color:", ";svg{fill:", ";}}"], _typographyConfig.default.fontStyles.regular, (0, _typography.pxToRem)(8), function (_ref) {
+var LeftArrowContainerStyles = _styledComponents.default.div.withConfig({
+  displayName: "prev__LeftArrowContainerStyles",
+  componentId: "sc-11dzafq-0"
+})(["color:#fff;", ";align-items:center;display:flex;transition:transform 0.3s ease-in-out;&:hover{transform:translate(-", ");cursor:pointer;color:", ";svg{fill:", ";}}"], _typographyConfig.default.fontStyles.regular, (0, _typography.pxToRem)(8), function (_ref) {
   var hoverColor = _ref.hoverColor;
   return hoverColor;
 }, function (_ref2) {
@@ -31,19 +31,20 @@ var RightArrowContainerStyles = _styledComponents.default.div.withConfig({
 });
 
 var TextStyles = _styledComponents.default.div.withConfig({
-  displayName: "next__TextStyles",
-  componentId: "bcawu0-1"
-})(["margin-right:", ";"], (0, _typography.pxToRem)(6));
+  displayName: "prev__TextStyles",
+  componentId: "sc-11dzafq-1"
+})(["margin-left:", ";"], (0, _typography.pxToRem)(6));
 
-var Next = function Next(_ref3) {
+var Prev = function Prev(_ref3) {
   var children = _ref3.children,
       hoverColor = _ref3.hoverColor;
-  return _react.default.createElement(RightArrowContainerStyles, {
+  return _react.default.createElement(LeftArrowContainerStyles, {
     hoverColor: hoverColor
-  }, _react.default.createElement(TextStyles, null, children), _react.default.createElement(_arrowIcon.ArrowIcon, {
+  }, _react.default.createElement(_arrowIcon.ArrowIcon, {
     version: "solid",
+    direction: "left",
     fill: "#fff"
-  }));
+  }), _react.default.createElement(TextStyles, null, children));
 };
 
-exports.Next = Next;
+exports.Prev = Prev;
