+++ src/components/carousel/tab-bar-carousel/index.js
@@ -1,7 +1,5 @@
 "use strict";
 
-var _interopRequireWildcard = require("@babel/runtime/helpers/interopRequireWildcard");
-
 var _interopRequireDefault = require("@babel/runtime/helpers/interopRequireDefault");
 
 Object.defineProperty(exports, "__esModule", {
@@ -9,72 +7,17 @@ Object.defineProperty(exports, "__esModule", {
 });
 exports.NavTabCarousel = void 0;
 
-var _classCallCheck2 = _interopRequireDefault(require("@babel/runtime/helpers/classCallCheck"));
+var _react = _interopRequireDefault(require("react"));
 
-var _createClass2 = _interopRequireDefault(require("@babel/runtime/helpers/createClass"));
+var _navTabCarousel = require("carousel/nav-tab-carousel");
 
-var _possibleConstructorReturn2 = _interopRequireDefault(require("@babel/runtime/helpers/possibleConstructorReturn"));
+var _deprecated = require("../../helpers/utils/deprecated");
 
-var _getPrototypeOf2 = _interopRequireDefault(require("@babel/runtime/helpers/getPrototypeOf"));
-
-var _inherits2 = _interopRequireDefault(require("@babel/runtime/helpers/inherits"));
-
-var _react = _interopRequireWildcard(require("react"));
-
-var _styledComponents = _interopRequireWildcard(require("styled-components"));
-
-var _propTypes = _interopRequireDefault(require("prop-types"));
-
-var _prev = require("./components/prev");
-
-var _next = require("./components/next");
-
-var _pagination = require("./components/pagination");
-
-var _colorsConfig = _interopRequireDefault(require("./../../../constants/colors-config"));
-
-var _reactElements = require("./../../../helpers/react/react-elements");
-
-var _typography = require("./../../../helpers/utils/typography");
-
-var NavTabCarouselStyles = _styledComponents.default.div.withConfig({
-  displayName: "tab-bar-carousel__NavTabCarouselStyles",
-  componentId: "sc-1vfblk9-0"
-})(["display:flex;width:100%;justify-content:space-between;align-items:center;background-color:", ";padding:", " ", ";"], _colorsConfig.default.font1, (0, _typography.pxToRem)(32), (0, _typography.pxToRem)(40));
-
-var NavTabCarousel =
-/*#__PURE__*/
-function (_Component) {
-  (0, _inherits2.default)(NavTabCarousel, _Component);
-
-  function NavTabCarousel() {
-    (0, _classCallCheck2.default)(this, NavTabCarousel);
-    return (0, _possibleConstructorReturn2.default)(this, (0, _getPrototypeOf2.default)(NavTabCarousel).apply(this, arguments));
-  }
-
-  (0, _createClass2.default)(NavTabCarousel, [{
-    key: "render",
-    value: function render() {
-      var children = this.props.children;
-      var prev = (0, _reactElements.getReactElementsByType)({
-        children: children,
-        type: NavTabCarousel.Prev
-      })[0];
-      var next = (0, _reactElements.getReactElementsByType)({
-        children: children,
-        type: NavTabCarousel.Next
-      })[0];
-      var pagination = (0, _reactElements.getReactElementsByType)({
-        children: children,
-        type: NavTabCarousel.Pagination
-      })[0];
-      return _react.default.createElement(NavTabCarouselStyles, null, prev, pagination, next);
-    }
-  }]);
-  return NavTabCarousel;
-}(_react.Component);
+var NavTabCarousel = function NavTabCarousel(props) {
+  var message = "DEPRECATED:\n    Import `NavTabCarousel` from `\u2026/nav-tab-carousel`\n    instead of `\u2026/tab-bar-carousel.";
+  return _react.default.createElement(_deprecated.Deprecated, {
+    warningMessage: message
+  }, _react.default.createElement(_navTabCarousel.NavTabCarousel, props));
+};
 
 exports.NavTabCarousel = NavTabCarousel;
-NavTabCarousel.Prev = _prev.Prev;
-NavTabCarousel.Next = _next.Next;
-NavTabCarousel.Pagination = _pagination.Pagination;
